+++ links/src/androidLibMain/kotlin/manager/PlatformLinksHandler.kt
@@ -1,0 +1,34 @@
+/*
+ Copyright 2023 Splendo Consulting B.V. The Netherlands
+
+    Licensed under the Apache License, Version 2.0 (the "License");
+    you may not use this file except in compliance with the License.
+    You may obtain a copy of the License at
+
+      http://www.apache.org/licenses/LICENSE-2.0
+
+    Unless required by applicable law or agreed to in writing, software
+    distributed under the License is distributed on an "AS IS" BASIS,
+    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+    See the License for the specific language governing permissions and
+    limitations under the License.
+
+ */
+
+package com.splendo.kaluga.links.manager
+
+import android.net.UrlQuerySanitizer
+import android.webkit.URLUtil
+
+actual class PlatformLinksHandler : LinksHandler {
+    override fun isValid(url: String): Boolean = URLUtil.isValidUrl(url)
+
+    override fun extractQueryAsList(url: String): List<Any> {
+        val params = UrlQuerySanitizer(url)
+
+        if (params.parameterList == null) {
+            return emptyList()
+        }
+        return params.parameterList.map { it.mValue }
+    }
+}
