+++ links/src/commonTest/kotlin/manager/MockLinksManager.kt
