--- links/src/jvmMain/kotlin/com/splendo/kaluga/links/manager/PlatformLinksHandler.kt
@@ -1,28 +1,0 @@
-/*
- Copyright 2022 Splendo Consulting B.V. The Netherlands
-
-    Licensed under the Apache License, Version 2.0 (the "License");
-    you may not use this file except in compliance with the License.
-    You may obtain a copy of the License at
-
-      http://www.apache.org/licenses/LICENSE-2.0
-
-    Unless required by applicable law or agreed to in writing, software
-    distributed under the License is distributed on an "AS IS" BASIS,
-    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-    See the License for the specific language governing permissions and
-    limitations under the License.
-
- */
-
-package com.splendo.kaluga.links.manager
-
-actual class PlatformLinksHandler : LinksHandler {
-    override fun isValid(url: String): Boolean {
-        TODO("Not yet implemented")
-    }
-
-    override fun extractQueryAsList(url: String): List<Any> {
-        TODO("Not yet implemented")
-    }
-}
